feat: adiciona modelo de curso com FK em aluno e corrige RM para BIGINT
- Cria tabela curso (nome, periodo, ativo) com FK em aluno.curso_id
- Adiciona endpoint GET /api/cursos (lista cursos ativos)
- Corrige RM de INT para BIGINT no schema (RM de 11 digitos estourava INT)
- Ajusta models Go de int/uint para int64/uint64 (aluno, transacao, historico, JWT)
- Adiciona Preload(Curso) em GetPerfilAluno e LoginAluno
- Cria utils/serie.go para calculo de serie esperada a partir do RM (uso futuro)
- Cria Makefile com atalhos para schema e seeds via Docker

+++ backend/controller/funcionarioController.go
@@ -38,7 +38,7 @@ func LoginFuncionario(c *gin.Context) {
 	}
 
 	//Gera o token JWT
-	token, err := utils.GenerateToken(funcionario.ID, "funcionario", 24*time.Hour)
+	token, err := utils.GenerateToken(uint64(funcionario.ID), "funcionario", 24*time.Hour)
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{
 			"error": "Erro ao gerar token JWT",
